checkpoint, tests pass thus far, but more need to be written
