+++ redis/connection.py
@@ -1,89 +1,29 @@
-from distutils.version import StrictVersion
+import os
+import socket
+import threading
+from contextlib import contextmanager
 from itertools import chain
 from time import time
 from queue import LifoQueue, Empty, Full
 from urllib.parse import parse_qs, unquote, urlparse
-import errno
-import io
-import os
-import socket
-import threading
-import warnings
 
-from redis.exceptions import (
+from .exceptions import (
     AuthenticationError,
     AuthenticationWrongNumberOfArgsError,
-    BusyLoadingError,
     ChildDeadlockedError,
     ConnectionError,
     DataError,
-    ExecAbortError,
-    InvalidResponse,
-    NoPermissionError,
-    NoScriptError,
-    ReadOnlyError,
     RedisError,
     ResponseError,
     TimeoutError,
-    ModuleError,
 )
-from redis.utils import str_if_bytes
-
-# attempt to import hiredis. if it is importable and new enough, flag that
-# hiredis support is enabled
-try:
-    import hiredis
-    HIREDIS_VERSION = StrictVersion(hiredis.__version__)
-    if HIREDIS_VERSION < StrictVersion('1.1.0'):
-        HIREDIS_AVAILABLE = False
-        msg = ('redis-py requires hiredis >= 1.1.0 to enable hiredis support. '
-               'You are running hiredis version %s. Please upgrade hiredis '
-               'to enable hiredis optimizations.') % HIREDIS_VERSION
-        warnings.warn(msg)
-    else:
-        HIREDIS_AVAILABLE = True
-
-except ImportError:
-    HIREDIS_AVAILABLE = False
-
-# attempt to import ssl. if it is importable, flag that ssl support is enabled
-try:
-    import ssl
-    SSL_AVAILABLE = True
-except ImportError:
-    SSL_AVAILABLE = False
-
-# sockets in nonblocking mode raise an exception when the socket would
-# otherwise block (e.g., trying to read from a socket that has no data
-# to be read). plain sockets raise BlockingIOError while sockets enabled with
-# ssl support raise various exceptions defined in the ssl module.
-# define a dict of exception class => error number that includes all
-# exceptions and error numbers raised when the socket would block but is
-# otherwise healthy.
-NONBLOCKING_EXCEPTION_ERROR_NUMBERS = {
-    BlockingIOError: errno.EWOULDBLOCK,
-}
+from .features import SSL_AVAILABLE
+from .parser import DefaultParser
+from .utils import DEFAULT, str_if_bytes
 
 if SSL_AVAILABLE:
-    if hasattr(ssl, 'SSLWantReadError'):
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLWantReadError] = 2
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLWantWriteError] = 2
-    else:
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLError] = 2
-
-NONBLOCKING_EXCEPTIONS = tuple(NONBLOCKING_EXCEPTION_ERROR_NUMBERS.keys())
-
-# constants used in redis protocol parsing
-SYM_STAR = b'*'
-SYM_DOLLAR = b'$'
-SYM_CRLF = b'\r\n'
-SYM_EMPTY = b''
-
-# the default value for keyword arguments where a None cannot be used
-SENTINEL = object()
-
-
-SERVER_CLOSED_CONNECTION_ERROR = "Connection closed by server."
+    # import the ssl module if it's available
+    import ssl
 
 
 class Encoder:
@@ -93,6 +33,32 @@ class Encoder:
         self.encoding = encoding
         self.encoding_errors = encoding_errors
         self.decode_responses = decode_responses
+
+    @contextmanager
+    def update(self, yielded_object, encoding=DEFAULT,
+               encoding_errors=DEFAULT, decode_responses=DEFAULT):
+        """
+        Update the encoding settings temporarily as a context manager
+
+        This is generally called from a method on an object that holds an
+        Encoder instance, e.g. the Connection or Client object.
+        """
+        prev_encoding = self.encoding
+        prev_encoding_errors = self.encoding_errors
+        prev_decode_responses = self.decode_responses
+
+        if encoding is not DEFAULT:
+            self.encoding = encoding
+        if encoding_errors is not DEFAULT:
+            self.encoding_errors = encoding_errors
+        if decode_responses is not DEFAULT:
+            self.decode_responses = decode_responses
+
+        yield yielded_object
+
+        self.encoding = prev_encoding
+        self.encoding_errors = prev_encoding_errors
+        self.decode_responses = prev_decode_responses
 
     def encode(self, value):
         "Return a bytestring or bytes-like representation of the value"
@@ -123,356 +89,6 @@ class Encoder:
         return value
 
 
-class BaseParser:
-    EXCEPTION_CLASSES = {
-        'ERR': {
-            'max number of clients reached': ConnectionError,
-            'Client sent AUTH, but no password is set': AuthenticationError,
-            'invalid password': AuthenticationError,
-            # some Redis server versions report invalid command syntax
-            # in lowercase
-            'wrong number of arguments for \'auth\' command':
-                AuthenticationWrongNumberOfArgsError,
-            # some Redis server versions report invalid command syntax
-            # in uppercase
-            'wrong number of arguments for \'AUTH\' command':
-                AuthenticationWrongNumberOfArgsError,
-            'Error loading the extension. Please check the server logs.':
-                ModuleError,
-            ('Error unloading module: the module exports one or more '
-             'module-side data types, can\'t unload'):
-                ModuleError,
-            'Error unloading module: no such module with that name':
-                ModuleError,
-            'Error unloading module: operation not possible.': ModuleError,
-        },
-        'EXECABORT': ExecAbortError,
-        'LOADING': BusyLoadingError,
-        'NOSCRIPT': NoScriptError,
-        'READONLY': ReadOnlyError,
-        'NOAUTH': AuthenticationError,
-        'NOPERM': NoPermissionError,
-    }
-
-    def parse_error(self, response):
-        "Parse an error response"
-        error_code = response.split(' ')[0]
-        if error_code in self.EXCEPTION_CLASSES:
-            response = response[len(error_code) + 1:]
-            exception_class = self.EXCEPTION_CLASSES[error_code]
-            if isinstance(exception_class, dict):
-                exception_class = exception_class.get(response, ResponseError)
-            return exception_class(response)
-        return ResponseError(response)
-
-
-class SocketBuffer:
-    def __init__(self, socket, socket_read_size, socket_timeout):
-        self._sock = socket
-        self.socket_read_size = socket_read_size
-        self.socket_timeout = socket_timeout
-        self._buffer = io.BytesIO()
-        # number of bytes written to the buffer from the socket
-        self.bytes_written = 0
-        # number of bytes read from the buffer
-        self.bytes_read = 0
-
-    @property
-    def length(self):
-        return self.bytes_written - self.bytes_read
-
-    def _read_from_socket(self, length=None, timeout=SENTINEL,
-                          raise_on_timeout=True):
-        sock = self._sock
-        socket_read_size = self.socket_read_size
-        buf = self._buffer
-        buf.seek(self.bytes_written)
-        marker = 0
-        custom_timeout = timeout is not SENTINEL
-
-        try:
-            if custom_timeout:
-                sock.settimeout(timeout)
-            while True:
-                data = self._sock.recv(socket_read_size)
-                # an empty string indicates the server shutdown the socket
-                if isinstance(data, bytes) and len(data) == 0:
-                    raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-                buf.write(data)
-                data_length = len(data)
-                self.bytes_written += data_length
-                marker += data_length
-
-                if length is not None and length > marker:
-                    continue
-                return True
-        except socket.timeout:
-            if raise_on_timeout:
-                raise TimeoutError("Timeout reading from socket")
-            return False
-        except NONBLOCKING_EXCEPTIONS as ex:
-            # if we're in nonblocking mode and the recv raises a
-            # blocking error, simply return False indicating that
-            # there's no data to be read. otherwise raise the
-            # original exception.
-            allowed = NONBLOCKING_EXCEPTION_ERROR_NUMBERS.get(ex.__class__, -1)
-            if not raise_on_timeout and ex.errno == allowed:
-                return False
-            raise ConnectionError("Error while reading from socket: %s" %
-                                  (ex.args,))
-        finally:
-            if custom_timeout:
-                sock.settimeout(self.socket_timeout)
-
-    def can_read(self, timeout):
-        return bool(self.length) or \
-            self._read_from_socket(timeout=timeout,
-                                   raise_on_timeout=False)
-
-    def read(self, length):
-        length = length + 2  # make sure to read the \r\n terminator
-        # make sure we've read enough data from the socket
-        if length > self.length:
-            self._read_from_socket(length - self.length)
-
-        self._buffer.seek(self.bytes_read)
-        data = self._buffer.read(length)
-        self.bytes_read += len(data)
-
-        # purge the buffer when we've consumed it all so it doesn't
-        # grow forever
-        if self.bytes_read == self.bytes_written:
-            self.purge()
-
-        return data[:-2]
-
-    def readline(self):
-        buf = self._buffer
-        buf.seek(self.bytes_read)
-        data = buf.readline()
-        while not data.endswith(SYM_CRLF):
-            # there's more data in the socket that we need
-            self._read_from_socket()
-            buf.seek(self.bytes_read)
-            data = buf.readline()
-
-        self.bytes_read += len(data)
-
-        # purge the buffer when we've consumed it all so it doesn't
-        # grow forever
-        if self.bytes_read == self.bytes_written:
-            self.purge()
-
-        return data[:-2]
-
-    def purge(self):
-        self._buffer.seek(0)
-        self._buffer.truncate()
-        self.bytes_written = 0
-        self.bytes_read = 0
-
-    def close(self):
-        try:
-            self.purge()
-            self._buffer.close()
-        except Exception:
-            # issue #633 suggests the purge/close somehow raised a
-            # BadFileDescriptor error. Perhaps the client ran out of
-            # memory or something else? It's probably OK to ignore
-            # any error being raised from purge/close since we're
-            # removing the reference to the instance below.
-            pass
-        self._buffer = None
-        self._sock = None
-
-
-class PythonParser(BaseParser):
-    "Plain Python parsing class"
-    def __init__(self, socket_read_size):
-        self.socket_read_size = socket_read_size
-        self.encoder = None
-        self._sock = None
-        self._buffer = None
-
-    def __del__(self):
-        try:
-            self.on_disconnect()
-        except Exception:
-            pass
-
-    def on_connect(self, connection):
-        "Called when the socket connects"
-        self._sock = connection._sock
-        self._buffer = SocketBuffer(self._sock,
-                                    self.socket_read_size,
-                                    connection.socket_timeout)
-        self.encoder = connection.encoder
-
-    def on_disconnect(self):
-        "Called when the socket disconnects"
-        self._sock = None
-        if self._buffer is not None:
-            self._buffer.close()
-            self._buffer = None
-        self.encoder = None
-
-    def can_read(self, timeout):
-        return self._buffer and self._buffer.can_read(timeout)
-
-    def read_response(self):
-        raw = self._buffer.readline()
-        if not raw:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        byte, response = raw[:1], raw[1:]
-
-        if byte not in (b'-', b'+', b':', b'$', b'*'):
-            raise InvalidResponse("Protocol Error: %r" % raw)
-
-        # server returned an error
-        if byte == b'-':
-            response = response.decode('utf-8', errors='replace')
-            error = self.parse_error(response)
-            # if the error is a ConnectionError, raise immediately so the user
-            # is notified
-            if isinstance(error, ConnectionError):
-                raise error
-            # otherwise, we're dealing with a ResponseError that might belong
-            # inside a pipeline response. the connection's read_response()
-            # and/or the pipeline's execute() will raise this error if
-            # necessary, so just return the exception instance here.
-            return error
-        # single value
-        elif byte == b'+':
-            pass
-        # int value
-        elif byte == b':':
-            response = int(response)
-        # bulk response
-        elif byte == b'$':
-            length = int(response)
-            if length == -1:
-                return None
-            response = self._buffer.read(length)
-        # multi-bulk response
-        elif byte == b'*':
-            length = int(response)
-            if length == -1:
-                return None
-            response = [self.read_response() for i in range(length)]
-        if isinstance(response, bytes):
-            response = self.encoder.decode(response)
-        return response
-
-
-class HiredisParser(BaseParser):
-    "Parser class for connections using Hiredis"
-    def __init__(self, socket_read_size):
-        if not HIREDIS_AVAILABLE:
-            raise RedisError("Hiredis is not installed")
-        self.socket_read_size = socket_read_size
-        self._buffer = bytearray(socket_read_size)
-
-    def __del__(self):
-        try:
-            self.on_disconnect()
-        except Exception:
-            pass
-
-    def on_connect(self, connection):
-        self._sock = connection._sock
-        self._socket_timeout = connection.socket_timeout
-        kwargs = {
-            'protocolError': InvalidResponse,
-            'replyError': self.parse_error,
-            'errors': connection.encoder.encoding_errors,
-        }
-
-        if connection.encoder.decode_responses:
-            kwargs['encoding'] = connection.encoder.encoding
-
-        self._reader = hiredis.Reader(**kwargs)
-        self._next_response = False
-
-    def on_disconnect(self):
-        self._sock = None
-        self._reader = None
-        self._next_response = False
-
-    def can_read(self, timeout):
-        if not self._reader:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        if self._next_response is False:
-            self._next_response = self._reader.gets()
-            if self._next_response is False:
-                return self.read_from_socket(timeout=timeout,
-                                             raise_on_timeout=False)
-        return True
-
-    def read_from_socket(self, timeout=SENTINEL, raise_on_timeout=True):
-        sock = self._sock
-        custom_timeout = timeout is not SENTINEL
-        try:
-            if custom_timeout:
-                sock.settimeout(timeout)
-            bufflen = self._sock.recv_into(self._buffer)
-            if bufflen == 0:
-                raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-            self._reader.feed(self._buffer, 0, bufflen)
-            # data was read from the socket and added to the buffer.
-            # return True to indicate that data was read.
-            return True
-        except socket.timeout:
-            if raise_on_timeout:
-                raise TimeoutError("Timeout reading from socket")
-            return False
-        except NONBLOCKING_EXCEPTIONS as ex:
-            # if we're in nonblocking mode and the recv raises a
-            # blocking error, simply return False indicating that
-            # there's no data to be read. otherwise raise the
-            # original exception.
-            allowed = NONBLOCKING_EXCEPTION_ERROR_NUMBERS.get(ex.__class__, -1)
-            if not raise_on_timeout and ex.errno == allowed:
-                return False
-            raise ConnectionError("Error while reading from socket: %s" %
-                                  (ex.args,))
-        finally:
-            if custom_timeout:
-                sock.settimeout(self._socket_timeout)
-
-    def read_response(self):
-        if not self._reader:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        # _next_response might be cached from a can_read() call
-        if self._next_response is not False:
-            response = self._next_response
-            self._next_response = False
-            return response
-
-        response = self._reader.gets()
-        while response is False:
-            self.read_from_socket()
-            response = self._reader.gets()
-
-        # if the response is a ConnectionError or the response is a list and
-        # the first item is a ConnectionError, raise it as something bad
-        # happened
-        if isinstance(response, ConnectionError):
-            raise response
-        elif isinstance(response, list) and response and \
-                isinstance(response[0], ConnectionError):
-            raise response[0]
-        return response
-
-
-if HIREDIS_AVAILABLE:
-    DefaultParser = HiredisParser
-else:
-    DefaultParser = PythonParser
-
-
 class Connection:
     "Manages TCP communication to and from a Redis server"
 
@@ -500,7 +116,8 @@ class Connection:
         self.next_health_check = 0
         self.encoder = Encoder(encoding, encoding_errors, decode_responses)
         self._sock = None
-        self._parser = parser_class(socket_read_size=socket_read_size)
+        self._parser = parser_class(encoder=self.encoder,
+                                    socket_read_size=socket_read_size)
         self._connect_callbacks = []
         self._buffer_cutoff = 6000
 
@@ -701,15 +318,13 @@ class Connection:
 
     def send_command(self, *args, **kwargs):
         "Pack and send a command to the Redis server"
-        self.send_packed_command(self.pack_command(*args),
+        self.send_packed_command(self._parser.pack_command(*args),
                                  check_health=kwargs.get('check_health', True))
 
     def can_read(self, timeout=0):
         "Poll the socket to see if there's data that can be read."
-        sock = self._sock
-        if not sock:
+        if not self._sock:
             self.connect()
-            sock = self._sock
         return self._parser.can_read(timeout)
 
     def read_response(self):
@@ -721,7 +336,7 @@ class Connection:
             raise TimeoutError("Timeout reading from %s:%s" %
                                (self.host, self.port))
         except socket.error as e:
-            self.disconnect()
+            self.odisconnect()
             raise ConnectionError("Error while reading from %s:%s : %s" %
                                   (self.host, self.port, e.args))
         except BaseException:
@@ -735,65 +350,24 @@ class Connection:
             raise response
         return response
 
-    def pack_command(self, *args):
-        "Pack a series of arguments into the Redis protocol"
-        output = []
-        # the client might have included 1 or more literal arguments in
-        # the command name, e.g., 'CONFIG GET'. The Redis server expects these
-        # arguments to be sent separately, so split the first argument
-        # manually. These arguments should be bytestrings so that they are
-        # not encoded.
-        if isinstance(args[0], str):
-            args = tuple(args[0].encode().split()) + args[1:]
-        elif b' ' in args[0]:
-            args = tuple(args[0].split()) + args[1:]
+    def execute(self, *args, return_response=True):
+        self.send_command(*args)
+        if return_response:
+            return self.read_response()
 
-        buff = SYM_EMPTY.join((SYM_STAR, str(len(args)).encode(), SYM_CRLF))
+    def encoding_context(self, encoding=DEFAULT,
+                         encoding_errors=DEFAULT, decode_responses=DEFAULT):
+        return self.encoder.update(self,
+                                   encoding=encoding,
+                                   encoding_errors=encoding_errors,
+                                   decode_responses=decode_responses)
 
-        buffer_cutoff = self._buffer_cutoff
-        for arg in map(self.encoder.encode, args):
-            # to avoid large string mallocs, chunk the command into the
-            # output list if we're sending large values or memoryviews
-            arg_length = len(arg)
-            if (len(buff) > buffer_cutoff or arg_length > buffer_cutoff
-                    or isinstance(arg, memoryview)):
-                buff = SYM_EMPTY.join(
-                    (buff, SYM_DOLLAR, str(arg_length).encode(), SYM_CRLF))
-                output.append(buff)
-                output.append(arg)
-                buff = SYM_CRLF
-            else:
-                buff = SYM_EMPTY.join(
-                    (buff, SYM_DOLLAR, str(arg_length).encode(),
-                     SYM_CRLF, arg, SYM_CRLF))
-        output.append(buff)
-        return output
+    # TODO: remove these two once refactor is complete
+    def pack_commands(self, *args, **kwargs):
+        return self._parser.pack_commands(*args, **kwargs)
 
-    def pack_commands(self, commands):
-        "Pack multiple commands into the Redis protocol"
-        output = []
-        pieces = []
-        buffer_length = 0
-        buffer_cutoff = self._buffer_cutoff
-
-        for cmd in commands:
-            for chunk in self.pack_command(*cmd):
-                chunklen = len(chunk)
-                if (buffer_length > buffer_cutoff or chunklen > buffer_cutoff
-                        or isinstance(chunk, memoryview)):
-                    output.append(SYM_EMPTY.join(pieces))
-                    buffer_length = 0
-                    pieces = []
-
-                if chunklen > buffer_cutoff or isinstance(chunk, memoryview):
-                    output.append(chunk)
-                else:
-                    pieces.append(chunk)
-                    buffer_length += chunklen
-
-        if pieces:
-            output.append(SYM_EMPTY.join(pieces))
-        return output
+    def pack_command(self, *args, **kwargs):
+        return self._parser.pack_command(*args, **kwargs)
 
 
 class SSLConnection(Connection):
@@ -859,7 +433,8 @@ class UnixDomainSocketConnection(Connection):
         self.next_health_check = 0
         self.encoder = Encoder(encoding, encoding_errors, decode_responses)
         self._sock = None
-        self._parser = parser_class(socket_read_size=socket_read_size)
+        self._parser = parser_class(encoder=self.encoder,
+                                    socket_read_size=socket_read_size)
         self._connect_callbacks = []
         self._buffer_cutoff = 6000
